preproc: generalize the loop handling, factor end of expansion handling
Combine together the end of mmacro and end of %rep processing, and
don't hardcode that any mmacro without a name is a %rep. Instead keep
a pointer to the level to which we need to unwind (we can only unwind
one level at a time, almost by definition) if any; if this pointer is
set we are currently unwinding.

We really shouldn't need three stacks (cond, mstk, istk) to do the job
of one, though.

Signed-off-by: H. Peter Anvin <hpa@zytor.com>