outlib: add infrastructure for common section and symbol handling
Pretty much all the backends have to do the same gymnastics to handle
symbols and sections. In the future, this should be done by the
assembly core, but in order to prepare the ground and get the
performance benefits as soon as possible, implement a library of
functions which can manage symbols and sections and their respective
hashes, trees, and tables.

This infrastructure is not yet used by any backend, that porting work
has to come next.

Signed-off-by: H. Peter Anvin (Intel) <hpa@zytor.com>